requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
python-multipart
httpx
python-dotenv
requests
email-validator
supabase 